Fixed bug introduced in merge

+++ src/main/java/net/mcreator/ui/modgui/PlantGUI.java
@@ -673,7 +673,7 @@ public class PlantGUI extends ModElementGUI<Plant> {
 		plant.growapableMaxHeight = (int) growapableMaxHeight.getValue();
 		plant.hardness = (double) hardness.getValue();
 		plant.resistance = (double) resistance.getValue();
-		plant.luminance = (double) luminance.getValue();
+		plant.luminance = (int) luminance.getValue();
 		plant.unbreakable = unbreakable.isSelected();
 		plant.forceTicking = forceTicking.isSelected();
 		plant.hasTileEntity = hasTileEntity.isSelected();
